Refactor: PD-246 비밀번호 필드 분리
