+++ src/features/sign-up/index.ts
@@ -1,1 +1,2 @@
 export { Email } from './ui/field/email'
+export { Password } from './ui/field/password'
